+++ docs/release-notes.md
@@ -14,7 +14,9 @@ meta:
 
 ### What's new
 
-* **Protocols**. [The Oasis Sapphire](/operations/oasis-sapphire/networks) support for elastic and dedicated nodes.
+* **Protocols**.
+   * [The Oasis Sapphire](/operations/oasis-sapphire/networks) support for elastic and dedicated nodes.
+   * [Polygon zkEVM Mainnet](/operations/polygon-zkevm/networks) support.
 * **Documentation**.
    * [Oasis Sapphire](/operations/oasis-sapphire).
    * [Oasis Sapphire tutorial](/tutorials/oasis-sapphire).